+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 * MISC
   * Intune: changed the display or instances to show the display name instead of Id during extraction.
   * M365DSCUtil: Fixed an issue on function Get-M365DSCExportContentForResource if ConnectionMode was set to anything but "Credentials*"
+  * Assert-M365DSCBlueprint: Add support to exclude resources from being compared
 
 # 1.23.426.3